Crate ocl_extras

Re-exports

pub use self::sub_buffer_pool::SubBufferPool;

pub use self::command_graph::CommandGraph;

pub use self::command_graph::Command;

pub use self::command_graph::CommandDetails;

pub use self::command_graph::KernelArgBuffer;

pub use self::command_graph::RwCmdIdxs;

pub use self::work_pool::WorkPool;

Modules

command_graph
A command requisite-dependency graph.

sub_buffer_pool
A sub-buffer allocator.

work_pool

Functions

scrambled_vec
Returns a vector with length size containing random values in the (half-open) range [vals.0, vals.1).

shuffle
Shuffles the values in a vector using a single pass of Fisher-Yates with a weak (not cryptographically secure) random number generator.

shuffled_vec
Returns a vector with length size which is first filled with each integer value in the (inclusive) range [vals.0, vals.1]. If size is greater than the number of integers in the aforementioned range, the integers will repeat. After being filled with size values, the vector is shuffled and the order of its values is randomized.