Struct ocl_extras::command_graph::Command

pub struct Command { /* fields omitted */ }

Methods

impl Command

pub fn new(details: CommandDetails) -> Command

pub fn preceding_writers(
&self,
cmds: &HashMap<usize, RwCmdIdxs>
) -> BTreeSet<usize>

Returns a list of commands which both precede a command and which write to a block of memory which is read from by that command.

pub fn following_readers(
&self,
cmds: &HashMap<usize, RwCmdIdxs>
) -> BTreeSet<usize>

Returns a list of commands which both follow a command and which read from a block of memory which is written to by that command.

pub fn details(&self) -> &CommandDetails