Struct ocl_extras::sub_buffer_pool::SubBufferPool

pub struct SubBufferPool<T: OclPrm> { /* fields omitted */ }

A simple (linear search) sub-buffer allocator.

Methods

impl<T: OclPrm> SubBufferPool<T>

pub fn new(len: u32, default_queue: Queue) -> SubBufferPool<T>

Returns a new buffer pool.

pub fn alloc(&mut self, len: u32, flags: Option<MemFlags>) -> Result<usize, ()>

Allocates space for and creates a new sub-buffer then returns the buffer id which can be used to ::get or ::free it.

pub fn free(&mut self, buffer_id: usize) -> Result<(), usize>

Deallocates the buffer identified by buffer_id or returns it back in the event of a failure.

pub fn get(&self, buffer_id: usize) -> Option<&Buffer<T>>

Returns a reference to the sub-buffer identified by buffer_id.

pub fn get_mut(&mut self, buffer_id: usize) -> Option<&mut Buffer<T>>

Returns a mutable reference to the sub-buffer identified by buffer_id.

pub fn defrag(&mut self)

Defragments the buffer. Be sure to ::finish() any and all command queues you may be using before doing this.

All kernels with a buffer argument set to any of the sub-buffers in this pool will need to be created anew or arguments refreshed (use Kernel::arg_..._named when initializing arguments in order to change them later).

pub fn resize(&mut self, len: u32)

Shrinks or grows and defragments the main buffer. Invalidates all kernels referencing sub-buffers in this pool. See ::defrag for more information.