Struct ocl_extras::work_pool::WorkPool

pub struct WorkPool { /* fields omitted */ }

A general purpose work completion pool.

Contains elements of a single-threaded event loop and a thread pool.

Runs in and manages its own threads. Dropping the WorkPool will block the dropping thread until all submitted and spawned work is complete.

Methods

impl WorkPool

pub fn new(buffer_size: usize) -> Result<WorkPool, WorkPoolError>

Create a new, empty work pool.

pub fn complete<F>(&mut self, future: F) -> Result<(), WorkPoolError> where
F: Future<Item = (), Error = Never> + Send + 'static,

Submits a future which need only be polled to completion and that contains no intensive CPU work (including memcpy).